test for table and numbering didSplit rules
